pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "imagekitio>=5.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
